# modules/analytics/architecture/openapi/v1/api.yaml
openapi: 3.0.3
info:
  title: Analytics API (Design-Time Specification)
  version: 1.0.0
  description: |
    **Design-Time OpenAPI Specification** - Source of truth for Analytics API contract
    
    This specification defines what SHOULD be implemented. Compare with runtime OpenAPI to track implementation progress.
    
    ## Overview
    Hyperspot Analytics API for managing queries, datasources, templates, layouts, and dashboards.
    The Analytics API provides a complete platform for building data-driven dashboards and reports.
    It uses the GTS (Generic Type System) for type-safe entity management.
    
    ## Features
    - **GTS Registry**: Register and manage GTS types and instances
    - **OData v4**: Compatible query interface with filtering, sorting, pagination
    - **Full-text search**: Advanced filtering and text search
    - **Field projection**: Select specific fields with $select
    - **Partial updates**: JSON Patch (RFC 6902) for granular updates
    
    ## Patterns
    - **Tolerant Reader**: Service understands field semantics per operation (create/update/read)
    - **Read-only entities**: Entities provisioned via configuration files cannot be modified
    - **Strict typing**: Entity payloads validated against GTS type schemas
    
    ## Resources
    - [GTS Specification](../openspec/specs/domain/gts-types.md)
    - [Tolerant Reader Pattern](https://martinfowler.com/bliki/TolerantReader.html)
    - [GTS Type Definitions](../gts/types/)
servers:
...
security:
  - oauth2: []
...
paths:
...
  /gts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: |
          GTS identifier following format:
          gts.<vendor>.<package>.<namespace>.<type>.v<MAJOR>[.<MINOR>][~[chain]]
          
          Validation rules:
          - Each segment: alphanumeric + underscore, start with letter
          - Version: v<MAJOR> required, <MINOR> optional
          - Can be type identifier (ends with ~) or instance identifier
          - Optional chain for derived types or instances
        schema:
          type: string
          pattern: ^gts\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.v[0-9]+(\.[0-9]+)?(~.*)?$
        example: gts.hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1

    get:
      summary: Get GTS Entity by ID
      tags: [GTS Registry]
      security:
        - oauth2: [analytics:gts_read]
      parameters:
        - $ref: '#/components/parameters/ContextTenantId'
      description: |
        **Required Scope:** `analytics:gts_read`
        
        Retrieve a specific GTS type or instance by exact identifier
      responses:
        '200':
          description: GTS entity details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GTSEntity'
              example:
                id: gts.hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1
                type: gts.hypernetix.hyperspot.ax.query.v1~
                entity:
                  category: gts.hypernetix.hyperspot.ax.category.v1~hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1
                  name: Sales Data Query
                  description: Query for retrieving sales analytics data
                  api_endpoint: https://api.acme.com/analytics/sales
                  capabilities_id: gts.hypernetix.hyperspot.ax.query_capabilities.v1~acme.analytics._.sales_capabilities.v1
                  returns_schema_id: gts.hypernetix.hyperspot.ax.schema.v1~hypernetix.hyperspot.ax.query_returns.v1~acme.analytics._.sales_data.v1
                registered_at: "2024-01-15T10:30:00Z"
                updated_at: "2024-01-15T14:22:00Z"
                deleted_at: null
                tenant: 550e8400-e29b-41d4-a716-446655440000
                registered_by: user@example.com
                updated_by: user@example.com
                deleted_by: null
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    put:
      summary: Update GTS Entity (Full Replacement)
      tags: [GTS Registry]
      description: |
        Replace entire entity object. All fields must be provided.
        
        **Restrictions:**
        - Only entities registered via API can be updated
        - Read-only entities (from config files) return 403
        - Server-managed fields (id, type, timestamps) are ignored
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GTSUpdateRequest'
      responses:
        '200':
          description: Entity successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GTSEntity'
        '403':
          $ref: '#/components/responses/ReadOnlyEntity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    patch:
      summary: Partially Update GTS Entity
      tags: [GTS Registry]
      security:
        - oauth2: [analytics:gts_write]
      parameters:
        - $ref: '#/components/parameters/ContextTenantId'
      description: |
        **Required Scope:** `analytics:gts_write`
        
        Partially update entity using JSON Patch (RFC 6902).
        
        **Supported Operations:**
        - replace, add, remove, copy, move, test
        
        **Restrictions:**
        - Only /entity/* paths can be modified
        - Attempts to modify server-managed fields are rejected
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/JsonPatchOperation'
            example:
              - op: replace
                path: /entity/description
                value: Updated description
              - op: replace
                path: /entity/api_endpoint
                value: https://api.acme.com/v2/analytics/sales
      responses:
        '200':
          description: Entity successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GTSEntity'
              example:
                id: gts.hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1
                type: gts.hypernetix.hyperspot.ax.query.v1~
                entity:
                  category: gts.hypernetix.hyperspot.ax.category.v1~hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1
                  name: Sales Data Query
                  description: Updated description via PATCH
                  api_endpoint: https://api.acme.com/v2/analytics/sales
                  capabilities_id: gts.hypernetix.hyperspot.ax.query_capabilities.v1~acme.analytics._.sales_capabilities.v1
                  returns_schema_id: gts.hypernetix.hyperspot.ax.schema.v1~hypernetix.hyperspot.ax.query_returns.v1~acme.analytics._.sales_data.v1
                registered_at: "2024-01-15T10:30:00Z"
                updated_at: "2024-01-16T11:22:00Z"
                deleted_at: null
                tenant: 550e8400-e29b-41d4-a716-446655440000
                registered_by: user@example.com
                updated_by: user@example.com
                deleted_by: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/ReadOnlyEntity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      summary: Delete GTS Entity (Soft Delete)
      tags: [GTS Registry]
      security:
        - oauth2: [analytics:gts_write]
      parameters:
        - $ref: '#/components/parameters/ContextTenantId'
      description: |
        **Required Scope:** `analytics:gts_write`
        
        Soft-delete entity by setting deleted_at timestamp.
        Entity remains in database for audit purposes.
      responses:
        '204':
          description: Entity successfully deleted (No Content)
        '403':
          $ref: '#/components/responses/ReadOnlyEntity'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        Service-to-Service OAuth2 authentication using Client Credentials flow.
        Services authenticate with client_id and client_secret to obtain access token with specific scopes.
      flows:
        clientCredentials:
          tokenUrl: https://auth.hyperspot.io/oauth/token
          scopes:
            analytics:metadata_read: Read OData service metadata
            analytics:gts_read: Read GTS entities, schemas, and metadata
            analytics:gts_write: Create, update, and delete GTS entities
            analytics:gts_admin: Manage tenant enablement configuration
            analytics:query_read: Read query metadata and schemas
            analytics:query_execute: Execute queries and access results
            analytics:template_read: Download template bundles
            analytics:template_write: Upload and manage template bundles
  schemas:
...
    GTSEntityObject:
      type: object
      description: |
        Generic entity object that conforms to a GTS type.
        The actual structure depends on the GTS type identifier.
        Server validates against the corresponding GTS schema.
      additionalProperties: true
    JsonSchemaType:
      type: object
      description: Valid JSON Schema for GTS type registration
      required:
        - $schema
        - $id
        - type
      properties:
        $schema:
          type: string
          format: uri
          enum:
            - https://json-schema.org/draft/2020-12/schema
          example: https://json-schema.org/draft/2020-12/schema
        $id:
          type: string
          pattern: ^gts://gts\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.[a-z][a-z0-9_]*\.v[0-9]+(\.[0-9]+)?(~.*)?~$
          description: |
            GTS type identifier URI following format:
            gts://gts.<vendor>.<package>.<namespace>.<type>.v<MAJOR>[.<MINOR>][~[chain]]~
            
            Validation rules:
            - URI scheme: gts://
            - Each segment: alphanumeric + underscore, start with letter
            - Version: v<MAJOR> required, <MINOR> optional
            - MUST end with ~ (type identifiers end with ~)
            - Optional chain before final ~ for derived types
          example: gts://gts.hypernetix.hyperspot.ax.query.v1~acme.custom._.my_query.v1~
        type:
          type: string
        description:
          type: string
        properties:
          type: object
          additionalProperties: true
        required:
          type: array
          items:
            type: string
        additionalProperties:
          type: object

    # ============================================================================
    # Request/Response Schemas
    # ============================================================================
...
    GTSUpdateRequest:
      type: object
      description: |
        Request for full entity replacement (PUT).
        Entity must conform to its GTS type schema.
      required:
        - entity
      properties:
        entity:
          oneOf:
            - $ref: '#/components/schemas/GTSEntityObject'
            - $ref: '#/components/schemas/JsonSchemaType'
          description: Complete entity object (instance data or JSON Schema for types)
    GTSEntity:
      type: object
      description: GTS entity with metadata
      required:
        - id
        - type
        - entity
        - registered_at
        - tenant
      properties:
        id:
          type: string
          description: GTS identifier
          readOnly: true
          example: gts.hypernetix.hyperspot.ax.query.v1~acme.analytics._.sales.v1
        type:
          type: string
          description: GTS type identifier (derived from id)
          readOnly: true
          example: gts.hypernetix.hyperspot.ax.query.v1~
        entity:
          oneOf:
            - $ref: '#/components/schemas/GTSEntityObject'
            - $ref: '#/components/schemas/JsonSchemaType'
          description: Entity data (instance conforming to GTS type, or JSON Schema for type registration)
        registered_at:
          type: string
          format: date-time
          description: Registration timestamp
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Soft deletion timestamp
          readOnly: true
        tenant:
          type: string
          format: uuid
          description: Tenant identifier
          readOnly: true
        registered_by:
          type: string
          description: User who registered the entity
          readOnly: true
        updated_by:
          type: string
          description: User who last updated the entity
          readOnly: true
        deleted_by:
          type: string
          nullable: true
          description: User who deleted the entity
          readOnly: true
...
    JsonPatchOperation:
      type: object
      required:
        - op
        - path
      properties:
        op:
          type: string
          enum: [add, remove, replace, move, copy, test]
        path:
          type: string
          description: JSON Pointer path
          example: /entity/description
        value:
          description: Value for add, replace, test operations
        from:
          type: string
          description: Source path for move, copy operations
...
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
  parameters:
    ContextTenantId:
      name: context_tenant_id
      in: query
      required: false
      description: |
        Execute request on behalf of a specific tenant within the authorized tenant hierarchy.
        
        **Cross-Tenant Access:**
        - Service validates that `context_tenant_id` is within the tenant hierarchy from JWT token
        - If not provided, defaults to `tenant_id` from JWT
        - Enables parent tenants to perform operations on behalf of child tenants
        
        **Use Cases:**
        - Admin service managing resources for multiple tenants
        - Parent org accessing child tenant data
        - Cross-tenant reporting and analytics
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            unsupportedQuery:
              summary: Unsupported query operation
              value:
                type: "https://api.hyperspot.io/problems/unsupported-query"
                title: Unsupported Query Operation
                status: 400
                detail: "Filter on 'entity/custom_field' is not supported. Available indexed fields: [id, type, tenant, registered_at, entity/api_endpoint, entity/name]"
...
    ReadOnlyEntity:
      description: Read-Only Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.hyperspot.io/problems/read-only-entity"
            title: Read-Only Entity
            status: 403
            detail: Entity is read-only. It was provisioned through configuration files and cannot be modified via the API.
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        RateLimit-Policy:
          schema:
            type: string
          description: |
            Rate limit policy in format "limit;w=window" where window is in seconds.
            IETF draft: https://datatracker.ietf.org/doc/html/draft-ietf-httpapi-ratelimit-headers
          example: "100;w=60"
        RateLimit-Limit:
          schema:
            type: integer
          description: Maximum number of requests allowed in the current window
          example: 100
        RateLimit-Remaining:
          schema:
            type: integer
          description: Number of requests remaining in the current window
          example: 0
        RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp (seconds) when the rate limit window resets
          example: 1704553200
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before making another request
          example: 60
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.hyperspot.io/problems/rate-limit-exceeded"
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded for tenant. Maximum 100 requests per minute allowed. Please retry after 60 seconds.
            instance: "https://api.hyperspot.io/analytics/v1/queries/gts.hypernetix.hyperspot.ax.query.v1~acme.sales"
    ServiceUnavailable:
      description: Service Unavailable - Temporary service disruption
      headers:
        Retry-After:
          schema:
            type: integer
          description: Number of seconds to wait before retrying the request
          example: 30
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.hyperspot.io/problems/service-unavailable"
            title: Service Unavailable
            status: 503
            detail: Analytics service is temporarily unavailable due to maintenance. Please retry after 30 seconds.